Record ctime for urledger mark time.
- creating separate class for UnderreplicatedLedger
- adding static final value for UNASSIGNED_CTIME
